+++ tests/gateway/test_connection_manager.py
@@ -11,7 +11,6 @@ import pytest
 from unittest.mock import Mock
 
 from healthchain.gateway.clients.fhir.sync.connection import FHIRConnectionManager
-from healthchain.gateway.fhir.errors import FHIRConnectionError
 from healthchain.gateway.api.protocols import FHIRServerInterfaceProtocol
 
 
@@ -27,42 +26,6 @@ def mock_fhir_client():
     client = Mock(spec=FHIRServerInterfaceProtocol)
     client.base_url = "https://test.fhir.com/R4"
     return client
-
-
-@pytest.mark.parametrize(
-    "connection_string,should_succeed",
-    [
-        # Valid connection strings
-        (
-            "fhir://epic.org/api/FHIR/R4?client_id=test&client_secret=secret&token_url=https://epic.org/token",
-            True,
-        ),
-        (
-            "fhir://localhost:8080/fhir?client_id=local&client_secret=pass&token_url=http://localhost/token",
-            True,
-        ),
-        # Invalid connection strings
-        ("http://not-fhir.com/api", False),  # Wrong scheme
-        ("fhir://", False),  # Missing hostname
-        ("invalid-string", False),  # Not a URL
-    ],
-)
-def test_connection_manager_source_validation_and_parsing(
-    connection_manager, connection_string, should_succeed
-):
-    """FHIRConnectionManager validates connection strings and parses hostnames correctly."""
-    if should_succeed:
-        connection_manager.add_source("test_source", connection_string)
-        assert "test_source" in connection_manager.sources
-        assert "test_source" in connection_manager._connection_strings
-        assert (
-            connection_manager._connection_strings["test_source"] == connection_string
-        )
-    else:
-        with pytest.raises(
-            FHIRConnectionError, match="Failed to parse connection string"
-        ):
-            connection_manager.add_source("test_source", connection_string)
 
 
 def test_connection_manager_client_retrieval_and_default_selection(
@@ -149,76 +112,3 @@ def test_connection_manager_cleanup_all_sources(connection_manager):
 
     mock_client1.close.assert_called_once()
     mock_client2.close.assert_called_once()
-
-
-def test_connection_manager_handles_invalid_source_gracefully(connection_manager):
-    """FHIRConnectionManager handles requests for unknown sources gracefully."""
-    # Add one valid source
-    connection_manager.add_source(
-        "valid_source",
-        "fhir://valid.com/fhir?client_id=test&client_secret=secret&token_url=https://valid.com/token",
-    )
-
-    # Request client for non-existent source
-    with pytest.raises(ValueError, match="Unknown source: invalid_source"):
-        connection_manager.get_client("invalid_source")
-
-    # Verify valid source still works
-    connection_manager._create_server_from_connection_string = Mock(return_value=Mock())
-    client = connection_manager.get_client("valid_source")
-    assert client is not None
-
-
-def test_connection_manager_handles_connection_string_corruption():
-    """FHIRConnectionManager handles corrupted connection strings gracefully."""
-    manager = FHIRConnectionManager()
-
-    # Add valid source first
-    manager.add_source(
-        "valid",
-        "fhir://valid.com/fhir?client_id=test&client_secret=secret&token_url=https://valid.com/token",
-    )
-
-    # Manually corrupt the connection string (simulating memory corruption or external modification)
-    manager._connection_strings["valid"] = "corrupted_string_not_url"
-
-    # Mock client creation to raise error for corrupted string
-    def mock_create_with_error(connection_string):
-        if connection_string == "corrupted_string_not_url":
-            raise ValueError("Invalid connection string format")
-        return Mock()
-
-    manager._create_server_from_connection_string = mock_create_with_error
-
-    # Should propagate the error appropriately
-    with pytest.raises(ValueError, match="Invalid connection string format"):
-        manager.get_client("valid")
-
-
-def test_connection_manager_memory_cleanup_on_source_removal():
-    """FHIRConnectionManager properly cleans up memory when sources are removed."""
-    manager = FHIRConnectionManager()
-
-    # Add source
-    manager.add_source(
-        "temp_source",
-        "fhir://temp.com/fhir?client_id=test&client_secret=secret&token_url=https://temp.com/token",
-    )
-
-    # Verify source was added
-    assert "temp_source" in manager.sources
-    assert "temp_source" in manager._connection_strings
-
-    # Simulate source removal (if the API existed)
-    # Since the current implementation doesn't have remove_source,
-    # we test manual cleanup to verify memory management
-    del manager.sources["temp_source"]
-    del manager._connection_strings["temp_source"]
-
-    # Verify cleanup worked
-    assert "temp_source" not in manager.sources
-    assert "temp_source" not in manager._connection_strings
-
-    # Verify the source is truly gone
-    with pytest.raises(ValueError, match="Unknown source: temp_source"):
-        manager.get_client("temp_source")
