+++ tests/gateway/test_fhir_gateway.py
@@ -1,6 +1,6 @@
 import pytest
 from unittest.mock import Mock, patch
-from typing import Dict, Any, List
+from typing import Dict, Any
 
 from fhir.resources.patient import Patient
 from fhir.resources.bundle import Bundle
@@ -48,109 +48,6 @@ def fhir_gateway(mock_connection_manager):
 def test_patient():
     """Fixture providing a test Patient resource."""
     return Patient(id="123", active=True)
-
-
-def test_transform_handler_registration_with_correct_annotation(fhir_gateway):
-    """Transform handlers with correct return type annotations register successfully."""
-
-    @fhir_gateway.transform(Patient)
-    def transform_patient(id: str, source: str = None) -> Patient:
-        return Patient(id=id)
-
-    assert fhir_gateway._resource_handlers[Patient]["transform"] == transform_patient
-
-
-def test_transform_handler_validation_enforces_return_type_match(fhir_gateway):
-    """Transform handler registration validates return type matches decorator resource type."""
-    from fhir.resources.observation import Observation
-
-    with pytest.raises(
-        TypeError, match="return type .* doesn't match decorator resource type"
-    ):
-
-        @fhir_gateway.transform(Patient)
-        def invalid_handler(id: str) -> Observation:  # Wrong return type
-            return Observation()
-
-
-def test_aggregate_handler_registration_without_validation(fhir_gateway):
-    """Aggregate handlers register without return type validation."""
-
-    @fhir_gateway.aggregate(Patient)
-    def aggregate_patients(id: str = None, sources: List[str] = None):
-        return []
-
-    assert fhir_gateway._resource_handlers[Patient]["aggregate"] == aggregate_patients
-
-
-def test_handler_registration_creates_routes(fhir_gateway):
-    """Handler registration automatically creates corresponding API routes."""
-    initial_routes = len(fhir_gateway.routes)
-
-    @fhir_gateway.transform(Patient)
-    def transform_patient(id: str) -> Patient:
-        return Patient(id=id)
-
-    assert len(fhir_gateway.routes) == initial_routes + 1
-
-
-def test_empty_capability_statement_with_no_handlers(fhir_gateway):
-    """Gateway with no handlers generates minimal CapabilityStatement."""
-    capability = fhir_gateway.build_capability_statement()
-
-    assert capability.model_dump()["resourceType"] == "CapabilityStatement"
-    assert capability.status == "active"
-    assert capability.kind == "instance"
-    assert capability.fhirVersion == "4.0.1"
-
-
-def test_capability_statement_includes_registered_resources(fhir_gateway):
-    """CapabilityStatement includes resources with registered handlers."""
-    from fhir.resources.observation import Observation
-
-    @fhir_gateway.transform(Patient)
-    def transform_patient(id: str) -> Patient:
-        return Patient(id=id)
-
-    @fhir_gateway.aggregate(Observation)
-    def aggregate_observations(id: str = None) -> List[Observation]:
-        return []
-
-    capability = fhir_gateway.build_capability_statement()
-    resources = capability.rest[0].resource
-    resource_types = [r.type for r in resources]
-
-    assert "Patient" in resource_types
-    assert "Observation" in resource_types
-
-
-def test_gateway_status_structure(fhir_gateway):
-    """Gateway status contains required fields with correct structure."""
-    status = fhir_gateway.get_gateway_status()
-
-    assert status["gateway_type"] == "FHIRGateway"
-    assert status["status"] == "active"
-    assert isinstance(status["timestamp"], str)
-    assert isinstance(status["version"], str)
-
-
-def test_supported_operations_tracking(fhir_gateway):
-    """Gateway accurately tracks registered operations."""
-    initial_ops = fhir_gateway.get_gateway_status()["supported_operations"][
-        "endpoints"
-    ]["transform"]
-
-    @fhir_gateway.transform(Patient)
-    def transform_patient(id: str) -> Patient:
-        return Patient(id=id)
-
-    updated_status = fhir_gateway.get_gateway_status()
-
-    assert (
-        updated_status["supported_operations"]["endpoints"]["transform"]
-        == initial_ops + 1
-    )
-    assert "Patient" in updated_status["supported_operations"]["resources"]
 
 
 def test_read_operation_with_client_delegation(fhir_gateway, test_patient):
